+++ source/Application/Model/Actions.php
@@ -7,12 +7,8 @@
 
 namespace OxidEsales\EshopCommunity\Application\Model;
 
-use oxDb;
-use oxField;
-use oxRegistry;
-use oxUtilsUrl;
-use oxUtilsView;
-use oxUtilsFile;
+use OxidEsales\Eshop\Core\Registry;
+use OxidEsales\EshopCommunity\Application\Controller\FrontendController;
 
 /**
  * Article actions manager. Collects and keeps actions of chosen article.
@@ -202,9 +198,13 @@ class Actions extends \OxidEsales\Eshop\Core\Model\MultiLanguageModel
      */
     public function getLongDesc()
     {
-        /** @var \OxidEsales\Eshop\Core\UtilsView $oUtilsView */
-        $oUtilsView = \OxidEsales\Eshop\Core\Registry::getUtilsView();
-        return $oUtilsView->parseThroughSmarty($this->oxactions__oxlongdesc->getRawValue(), $this->getId() . $this->getLanguage(), null, true);
+        $activeView = oxNew(FrontendController::class);
+        $activeView->addGlobalParams();
+        $utilsView = Registry::getUtilsView();
+        return $utilsView->getRenderedContent(
+            $this->oxactions__oxlongdesc->getRawValue(),
+            $activeView->getViewData(),
+            $this->getId() . $this->getLanguage());
     }
 
     /**