+++ source/Application/Model/Article.php
@@ -13,6 +13,7 @@ use OxidEsales\Eshop\Core\Registry;
 use OxidEsales\Eshop\Core\Str;
 use OxidEsales\Eshop\Core\TableViewNameGenerator;
 use OxidEsales\EshopCommunity\Core\DatabaseProvider;
+use OxidEsales\EshopCommunity\Application\Controller\FrontendController;
 use oxList;
 
 // defining supported link types
@@ -2601,7 +2602,16 @@ class Article extends \OxidEsales\Eshop\Core\Model\MultiLanguageModel implements
      */
     public function getLongDesc()
     {
-        return Registry::getUtilsView()->parseThroughSmarty($this->getLongDescription()->getRawValue(), $this->getId() . $this->getLanguage(), null, true);
+        if ($this->getLongDescription() && $this->getLongDescription()->getRawValue()) {
+            $activeView = oxNew(FrontendController::class);
+            $activeView->addGlobalParams();
+            $utilsView = Registry::getUtilsView();
+            return $utilsView->getRenderedContent(
+                $this->getLongDescription()->getRawValue(),
+                $activeView->getViewData(),
+                $this->getId() . $this->getLanguage());
+        }
+        return '';
     }
 
     /**